Update tests to inspect the output more carefully
Why is this change needed?
--------------------------
We're introducing some new options to the clients, and it makes sense to have some more complete tests in place as we do

How does it address the issue?
------------------------------
Simply checks the output of valid-type tests to make sure that the resulting file is the expected type

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------