Update the swagger config to more correctly implement the hosted docs API.
Why is this change needed?
--------------------------
Adding commands to automatically download hosted docs resulted in a weird multi-endpoint mess that blurred the line between hosted and async.

How does it address the issue?
------------------------------
This makes the final result of a hosted (sync or async) document a URL, and assumes that the user will use whatever normal download-from-url tools are in their environment to retrieve the document as desired.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------