Adds a hosted async document example file
Why is this change needed?
--------------------------
The way hosted documents work is different that how DocRaptor has worked
in the past. Since we've added a new method to help create hosted async
documents we should add an example to portray how that works.

How does it address the issue?
------------------------------
This creates a new example file that creates and hosted async document,
downloads the document from the public url, and writes it to a temporary
file.

There are two key differences between this example and the other async
example:

1. Since it is a publicly available download we no longer have to make
   an authenticated request for the document.
2. We don't allow the example api key to create hosted documents so
   user's will need to provide their own if they want this to work.