Only create test documents in the test files
Why is this change needed?
--------------------------
Testing the the host name of a hosted document's download URL seemed
a little too coupled to DocRaptors API implementation. Since we only
care if our agent methods work we can use test documents across the
board. Test documents only stay hosted for a limited amount of time, but
they function the same as normal documents.

How does it address the issue?
------------------------------
This updates the test parameter to true on all tests and removes the
check against the download URL. This allows us to test documents without
having to worry so much about them sticking around for a long time.