Re-add ruby 2.5 support and restrict to >= 2.5
Why is this change needed?
--------------------------
This was 2.3, so this is also removing 2.3 and 2.4.

2.3 was removed because it's very old and the openapi-generated code is
>= 2.4 be default.

2.4 is removed because openapi-generator has a (so far) unreported
bug[1] because they started using the `transform_keys` method, which
was introduced in ruby 2.5, but they didn't update the gemspec to 2.5.

How does it address the issue?
------------------------------
Add ruby 2.5.

Update the test to handle old versions of `file`.

Update the docker script to mount an image specific cache dir. Without
this older rubies will try to use the same bundle which breaks on
compiled extensions that were compiled on a different ruby version.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
[1] - https://github.com/OpenAPITools/openapi-generator/commit/b1c6513015531f83418d393092a04293d424ded9
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?