Update gemspec fixer
Why is this change needed?
--------------------------
This script failed because the "bad" strings weren't found.

How does it address the issue?
------------------------------
Remove the unneccessary code. Switch to use the config for the ruby
version.

It's worth noting the speed improvement from using `git ls-files`.
Without this, every time you interact with anything bundler it will
parse the gemspec and therefore scan every file in your directory. If
your bundle is the the app dir so it can be cached between docker runs
this means there are a huge number of files. File system interaction is
slower inside docker, so this all adds up to very long test runs.
`git ls-files` results in a much smaller set of files to interact with,
and is much faster.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?