Clarify full test run output
Why is this change needed?
--------------------------
Previously it might not be clear if an individual test run passed or if
all tests runs had passed.

How does it address the issue?
------------------------------
Clarify that the full suite passed or failed at the end of the run.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?