Add facility for apply patches post generation
Why is this change needed?
--------------------------
Some agents use this, and it's nice to keep these script in sync.

How does it address the issue?
------------------------------
Allow a directory of patch files to be applied after generation.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?