Add a note about test improvements
Why is this change needed?
--------------------------
Ideally we would test against known working versions of all the
dependency gems, as well as updated versions of those gems. This would
help ensure the agent works for possible customer configurations.

How does it address the issue?
------------------------------
In order to keep scope down this is just a note for now.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?