Rename some swagger -> openapi references
Why is this change needed?
--------------------------
We want to show we're using a more maintained tool, and prevent
confusion for users and developers.

How does it address the issue?
------------------------------
Rename some easy and obvious references from swagger to openapi.

This also includes an update to the readme removing the
generate_language argument since it's not longer necessary.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?