Update version bump process
Why is this change needed?
--------------------------
We should rely on the generator to change the version number in all the
various generated locations instead of manually changing it. This will
be less error prone, especially if the locations of the generated
version numbers change.

How does it address the issue?
------------------------------
Switch to updating the version number in the generator config then
regenerating.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?