Update openapi-generator to 6.6.0
Why is this change needed?
--------------------------
This fixes an issue with setting the `host` in the client, which was
previously ignored. This was necessary for testing against the staging
environment.

How does it address the issue?
------------------------------
Update .generator-revision to 6.6.0 and regenerate the client.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298552

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?