Allow running the tests with an API key
Why is this change needed?
--------------------------
This is helpful to test options that regular users have access to,
instead of the more limited options available to the YOUR_API_KEY_HERE
user.

How does it address the issue?
------------------------------
If we can find a key in the file tmp/.docraptor_key, we'll rerun the
test with that key set in the environment. In either case we'll run the
test without the key.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298552

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?