Fix issue with failing tests not failing the run
Why is this change needed?
--------------------------
Tests that used `tmp/.docraptor_key` were failing, but the overall test
run was still passing. This was because the `run_test` function was not
returning a non-zero exit code when the test failed.

How does it address the issue?
------------------------------
Bubble up the return status of the test that uses the key file.

Since using the key file doubles the number of tests, I also added some
nicer colored output to the test script to make it easier to see which
tests are failing.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298552

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?