Add new API options
Why is this change needed?
--------------------------
We're planning on making these options available soon, so we'd like to
add them to the clients.

How does it address the issue?
------------------------------
Add the options and regenerate the client.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298552

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?