Refactor volume mounting be configurable via a file
Why is this change needed?
--------------------------
Different clients have different requirements. Ruby needs ssh to
publish a release because it needs to connect to github. This should
support whatever the client needs.

How does it address the issue?
------------------------------
This moves the magic ssh configuration into the file with the rest of
the volume info.

We need to jump through a few hoops because the cache dir has a colon in
it, and that's the field separator for docker `--volume` mounts. We need
`--volume` instead of `--mount` because the ssh magic path needs
`--volume`.

We do some eval so we can replace the colon with an underscore in the
volume list.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298552
https://gist.github.com/d11wtq/8699521

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?