Add a test for the new iframes option
Why is this change needed?
--------------------------
We want to verify the new options works for users.

How does it address the issue?
------------------------------
Test turning iframes off and check that we don't see the iframe content
in the output.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298552

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?