Update to the latest docraptor.yaml spec
Why is this change needed?
--------------------------
Whenever we update the spec we need to regenerate the client.

In this case we're updating to OAS 3.0.0.

How does it address the issue?
------------------------------
Update the spec and regenerate the client.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todos/6362751178

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?